Switch PINSFVVelocityVariable to SuperficialVelocityVariable
