+++ modules/navier_stokes/doc/content/source/variables/PINSFVSuperficialVelocityVariable.md
@@ -1,0 +1,11 @@
+# PINSFVSuperficialVelocityVariable
+
+This variable type is specific to the porous media incompressible Navier Stokes equations.
+Kernels for these equations will expect this variable type to be used, because they are
+written for superficial velocity non-linear variables.
+
+!syntax parameters /Adaptivity/Markers/PINSFVSuperficialVelocityVariable
+
+!syntax inputs /Adaptivity/Markers/PINSFVSuperficialVelocityVariable
+
+!syntax children /Adaptivity/Markers/PINSFVSuperficialVelocityVariable
