+++ modules/navier_stokes/src/variables/PINSFVSuperficialVelocityVariable.C
@@ -1,0 +1,23 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "PINSFVSuperficialVelocityVariable.h"
+
+registerMooseObject("NavierStokesApp", PINSFVSuperficialVelocityVariable);
+
+InputParameters
+PINSFVSuperficialVelocityVariable::validParams()
+{
+  return INSFVVelocityVariable::validParams();
+}
+
+PINSFVSuperficialVelocityVariable::PINSFVSuperficialVelocityVariable(const InputParameters & params)
+  : INSFVVelocityVariable(params)
+{
+}
