+++ modules/navier_stokes/test/tests/finite_volume/pins/channel-flow/2d-rc-epsjump.i
@@ -21,11 +21,11 @@ velocity_interp_method='rc'
 
 [Variables]
   [u]
-    type = PINSFVVelocityVariable
+    type = PINSFVSuperficialVelocityVariable
     initial_condition = 1
   []
   [v]
-    type = PINSFVVelocityVariable
+    type = PINSFVSuperficialVelocityVariable
     initial_condition = 1e-6
   []
   [pressure]
