+++ modules/navier_stokes/test/tests/finite_volume/pins/channel-flow/2d-rc.i
@@ -24,11 +24,11 @@ velocity_interp_method='rc'
 [Variables]
   inactive = 'lambda'
   [u]
-    type = PINSFVVelocityVariable
+    type = PINSFVSuperficialVelocityVariable
     initial_condition = 1
   []
   [v]
-    type = PINSFVVelocityVariable
+    type = PINSFVSuperficialVelocityVariable
     initial_condition = 1e-6
   []
   [pressure]
